ref: move OrderDirection enum
