+++ tests/Helper/QueryDataProvider.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Helper;
 
-use App\Core\Query\OrderDirection;
+use App\Core\Query\Enum\OrderDirection;
 use Illuminate\Foundation\Testing\WithFaker;
 
 class QueryDataProvider
